+++ kbox2d/src/commonMain/kotlin/org/jbox2d/dynamics/contacts/PositionSolverManifold.kt
@@ -9,10 +9,8 @@ import org.jbox2d.internal.*
 internal class PositionSolverManifold {
 
     val normal = Vec2()
-
     val point = Vec2()
-
-    var separation: Float = 0.toFloat()
+    var separation: Float = 0f
 
     fun initialize(pc: ContactPositionConstraint, xfA: Transform, xfB: Transform, index: Int) {
         assert(pc.pointCount > 0)
